+++ data-api/app/controllers/agents/__init__.py
@@ -1,13 +1,11 @@
-# TODO: remove BackgroundScheduler fron data_download_agent.py
-# TODO: remove BackgroundScheduler from contorllers/__init__.py
 from apscheduler.schedulers.background import BackgroundScheduler
 import threading
 from loguru import logger
-from app.controllers.sensitivity_analysis_agent import convert_data
-from app.controllers.data_downloader_agent import download_data
-from app.controllers.uncertainty_mean_sample_agent import uncertainty_mean_sample_agent
-from app.controllers.uncertainty_clustering_agent import uncertainty_clustering_agent
-from app.controllers.uncertainty_cluster_mean_sample_agent import uncertainty_cluster_mean_sample_agent
+from app.controllers.agents.sensitivity_analysis_agent import convert_data
+from app.controllers.agents.data_downloader_agent import download_data
+from app.controllers.agents.uncertainty_mean_sample_agent import uncertainty_mean_sample_agent
+from app.controllers.agents.uncertainty_clustering_agent import uncertainty_clustering_agent
+from app.controllers.agents.uncertainty_cluster_mean_sample_agent import uncertainty_cluster_mean_sample_agent
 
 
 def uncertainty_agents():
@@ -20,8 +18,9 @@ def uncertainty_agents():
     threading.Thread(target=uncertainty_mean_sample_agent).start()
 
     t_u_cluster.join()
-    print("Running Uncertainty Clustering Complete")
+    print("Uncertainty Clustering Complete")
     # Start computations requiring clustered data
+    print("Running Uncertainty Cluster Analysis")
     threading.Thread(target=uncertainty_cluster_mean_sample_agent).start()
 
 
@@ -33,6 +32,7 @@ def sensitivity_agents():
 def run_agents():
     logger.info('Download data agent starts. Will run immediately now and every midnight.')
     download_data()
+    threading.main_thread().run()  # Resume main thread
     threading.Thread(target=uncertainty_agents).start()
     threading.Thread(target=sensitivity_agents).start()
 
@@ -40,11 +40,10 @@ def run_agents():
 # A recurrent job
 scheduler = BackgroundScheduler(daemon=True)
 
-# Cron runs at 1am daily
+# Cron runs at 0am daily
 scheduler.add_job(run_agents, "cron", hour=0, minute=0, second=0)
 
 scheduler.start()
-logger.info('Uncertainty-clustering-agent starts. Will run immediately now and every 1am.')
 
 # Run immediately after server starts
-threading.Thread(target=run_agents).start()
+run_agents()
