+++ data-api/app/controllers/agents/uncertainty_cluster_mean_sample_agent.py
@@ -5,9 +5,9 @@ import json
 from loguru import logger
 from apscheduler.schedulers.background import BackgroundScheduler
 
-import app.controllers.uncertainty_analysis.clustering_tools as ct
-import app.controllers.uncertainty_analysis.uncertainty_mean_sample as ums
-import app.controllers.uncertainty_analysis.uncertainty_model_inventory as inventory
+import app.controllers.agents.uncertainty_analysis.clustering_tools as ct
+import app.controllers.agents.uncertainty_analysis.uncertainty_mean_sample as ums
+import app.controllers.agents.uncertainty_analysis.uncertainty_model_inventory as inventory
 
 from app.core.settings import DATA_PATH_LIVE
